Initial implementation of conversion_electrode.
This begins the work implementing a conversion_electrode model. This
represents an electrode that converts one solid phase to another, during
charge/disccharge.  The most prominent example is a Li-sulfur cathode.

The `__init__` and `intialize` routines are complete and functional, but
no other work has yet commenced.  Running the `Li_PorousSep_Sulfur.yaml`
therefore produces an error (the `residual` function contents are those
 from the single-particle-electrode.